# some work on deformed Burau matrix