@tscrim 2024-11-18
Do you have some timings?