@fchapoton 2024-11-18
before :
```
sage: ans = Knot([[1, 5, 2, 4], [3, 7, 4, 6], [5, 3, 6, 2], [7, 13, 8, 12], [9, 
....: 18, 10, 19], [11, 15, 12, 14], [13, 9, 14, 8], [15, 22, 16, 1], [17, 20, 1
....: 8, 21], [19, 10, 20, 11], [21, 16, 22, 17]])
sage: g = ans.braid()
sage: g.Tietze()
( 1, 2, 3, 3, 3, -2, -1, 3, 3, -2, -3, -4, 5, -4, 3, 2, -4, -4, 3, -5, -4 )
sage: %timeit g.deformed_burau_matrix()
35.6 ms ± 1.17 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

after :
```
sage: %timeit g.deformed_burau_matrix()
31.9 ms ± 87.7 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

So not very much faster. But I was not expecting more.